#64 Timeline.vueをcomponentsからviewsへ移動
